fix error if file not found

+++ src/webtail/Services/FileService.cs
@@ -1,8 +1,9 @@
 ﻿namespace webtail.Services
 {
+	using Microsoft.AspNetCore.Hosting;
+	using Radzen;
 	using System.IO;
 	using System.Threading.Tasks;
-	using Microsoft.AspNetCore.Hosting;
 
 	public class FileService
 	{
@@ -22,7 +23,10 @@
 			var randomName = $"{Guid.NewGuid()}{Path.GetExtension(sourceFile)}";
 			var destPath = Path.Combine(filesDir, randomName);
 			var sourcePath = Path.Combine(_env.WebRootPath, "files", sourceFile);
-
+			if (!File.Exists(sourcePath))
+			{ 
+				return "_";
+			}
 			await Task.Run(() => File.Copy(sourcePath, destPath, overwrite: true));
 
 			return randomName;
